+++ web/lib/agent/agent.client.ts
@@ -19,7 +19,6 @@ interface AgentClientOptions {
     serverEndpoint: string
     accessToken: string
     createAgentWorker: () => Worker
-    workspaceRootUri: string
     telemetryClientName?: string
     customHeaders?: Record<string, string>
     debug?: boolean
@@ -30,7 +29,6 @@ export async function createAgentClient({
     serverEndpoint,
     accessToken,
     createAgentWorker,
-    workspaceRootUri,
     customHeaders,
     telemetryClientName,
     debug = true,
@@ -70,8 +68,11 @@ export async function createAgentClient({
     const serverInfo: ServerInfo = await rpc.sendRequest('initialize', {
         name: process.env.CODY_WEB_DEMO ? 'standalone-web' : 'web',
         version: '0.0.1',
-        workspaceRootUri,
+        // Empty root URI leads to openctx configuration resolution failure, any non-empty
+        // mock value (Cody Web doesn't really use any workspace related features)
+        workspaceRootUri: 'sourcegraph/cody',
         capabilities: {
+            edit: 'none',
             completions: 'none',
             webview: 'agentic',
             disabledMentionsProviders: [FILE_CONTEXT_MENTION_PROVIDER.id],